+++ RELEASE-NOTES.md
@@ -11,9 +11,17 @@ VERSION SCHEMA: vYYYY.MM.#SEQ
 
 ---
 
-## v2020.02.1
+## v2020.03.2
+
+- NEW: Add new NIC logo.
+- FIX: Fix bug in crime types lookup.
+
+---
+
+## v2020.03.1
 
 - NEW: Miscreant Fingerprints Enrollment Workflow.
+- FIX: Fix bug in crime types lookup.
 
 ---